When a ticket is resolved. it's asking if the following. Action Required: Issue Resolved? This ticket has been marked as resolved. Please review the solution provided above.
If the solution addresses your issue, please click the button below to close your ticket.
If you are still experiencing problems, add a reply to this ticket (using the reply box above if available, or by creating a new one if not), and the ticket status will be automatically re-opened for further assistance. Right now the user cannot reply anymore to the ticket if it is marked as resolve. please let the user continue to reply to the resolved ticket until fully resolved. then the user can click the Yes close my ticket